+++ flink-kubernetes/src/main/java/org/apache/flink/kubernetes/kubeclient/Fabric8FlinkKubeClient.java
@@ -160,7 +160,7 @@ public class Fabric8FlinkKubeClient implements FlinkKubeClient {
 			deployment = d.decorate(deployment);
 		}
 
-		deployment = new FlinkMasterDeploymentDecorator(clusterSpecification).decorate(deployment);
+		deployment = new FlinkMasterDeploymentDecorator(clusterSpecification, internalClient).decorate(deployment);
 
 		LOG.debug("Create Flink Master deployment with spec: {}", deployment.getInternalResource().getSpec());
 
@@ -181,7 +181,7 @@ public class Fabric8FlinkKubeClient implements FlinkKubeClient {
 					pod = d.decorate(pod);
 				}
 
-				pod = new TaskManagerPodDecorator(parameter).decorate(pod);
+				pod = new TaskManagerPodDecorator(parameter, internalClient).decorate(pod);
 
 				LOG.debug("Create TaskManager pod with spec: {}", pod.getInternalResource().getSpec());